% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_functional_logger.R
\name{flog_filter_df}
\alias{flog_filter_df}
\title{flog_filter_df}
\usage{
flog_filter_df(.data, filter_dots, logger)
}
\arguments{
\item{.data}{A LoggingTuple or a raw dataset. The dataset
entry should be a data.frame. See drug.markers::flog.}

\item{filter_dots}{A vector or list of strings that define the
filters that shouold be applied to a data.frame; each of these should be a
valid input to the .dots arg of dplyr::filter_. If any of the entries is
"identity", then the data.frame that is passed in is unaltered by that
specific filtering step (although may be altered by subsequent steps).}

\item{logger}{A single logging function(post, pre) for
comparing a dataframe after a single filtering step (post) to the dataframe
that before that step (pre). See drug.markers::flog for more details. This
logging function is applied after each filtering step in filter_dots has
ran.}
}
\description{
Applies a sequence of filtering steps to a data.frame and returns a
(dataset, logdata) tuple where the logdata list contains a single
data.frame generated by combining the results of each individual logging
function
}

