% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_functional_logger.R
\docType{class}
\name{LoggingStep-class}
\alias{LoggingStep}
\alias{LoggingStep-class}
\title{Constructor for LoggingStep Class.}
\arguments{
\item{modifier}{A function that modifies a given dataset. This
must be able
to take the dataset as it's first parameter and have no other missing-but
necessary parameters}

\item{logger}{A function that can compare the outpur from
`modifier`
against the input to that function and return a (preferably small)
summary of the difference between output and input. It has obligatory
first parameter-name `post` (holds the dataset after the corresponding
`modifier` has ran) and may have an optional second parameter called `pre`
(which holds the dataset that was passed into the current `modifier`)}
}
\description{
Encapsulates a single data-manipulation function (`modifier`) and a single
  logging function (`logger`) to compare the input/output of `modifier`.
The user should not use the (non-exported) `run_step` method of LoggingStep
  directly. They should always use `flog(in.list = some.list,
  logsteps = LS)` or `flog(in.data = some.input, logsteps = LS)`.

Initialize method for LoggingStep
}

