% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_functional_logger.R
\docType{class}
\name{LoggingTuple-class}
\alias{LoggingTuple}
\alias{LoggingTuple-class}
\title{Constructor for LoggingTuple Class.}
\arguments{
\item{dataset}{May be any datatype, but must be defined.}

\item{logdata}{A list. If missing, this is set to the empty
list.}
}
\description{
Lightweight datastructure that is passed between different LoggingStep
functions. This holds the `current` dataset and a log of the various steps
that resulted in the generation of the current dataset.

Initializer for LoggingTuple
}

