% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_functional_logger.R
\name{nrow_logger}
\alias{nrow_logger}
\title{nrow_logger: Logs the change in the number of rows in a dataset after
  applying the modifier in a LoggingStep.}
\usage{
nrow_logger(post, pre)
}
\arguments{
\item{post}{The dataset that exists after a modifier has
been applied to some input dataset in a LoggingStep.}

\item{pre}{A dataset before a modifier has been applied to
it in a LoggingStep.}
}
\description{
nrow_logger: Logs the change in the number of rows in a dataset after
  applying the modifier in a LoggingStep.
}

