% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_functional_logger.R
\name{null_logger}
\alias{null_logger}
\title{null_logger: Logs no information about the changes induced by a modifier in
  a LoggingStep.}
\usage{
null_logger(post, pre)
}
\arguments{
\item{post}{The dataset that exists after a modifier has
been applied to some input dataset in a LoggingStep.}

\item{pre}{A dataset before a modifier has been applied to
it in a LoggingStep.}
}
\description{
null_logger: Logs no information about the changes induced by a modifier in
  a LoggingStep.
}

